feat: make it possible to make a guaranteed subscription to events
There was a problem that if for some reason
the service was not available or something with rabbitMQ,
There was a problem that if for some reason
the service was not available or something with rabbitMQ,
emails may not have been sent and the queue in RabbitMQ may have just cleared out.
It was decided to create a separate queue for notification-service
and link it to exchange from auth-service.
We also added a confirmation of message receipt in RabbitMQ
and resending the message in case of failure,
which also gives a guarantee that the user will be get messages via email

!!! Note:
For correct operation, you must manually configure queues
and connect them to exchange in the RabbitMQ interface